调整 租户 分页查询

+++ smart-common/smart-common-core/src/main/java/com/lee/common/core/Pagination.java
@@ -3,12 +3,12 @@ package com.lee.common.core;
 import lombok.Data;
 
 /**
- * 分页查询
+ * 分页信息
  * @author lee.li
  */
 @Data
 public class Pagination {
-    private Integer total;
-    private Integer current;
-    private Integer pageSize;
+    private Long total;
+    private Long current;
+    private Long pageSize;
 }
